fix: added suggestions
